+++ app/src/main/java/com/github/pockethub/android/ui/issue/EditMilestoneTask.java
@@ -15,20 +15,17 @@
  */
 package com.github.pockethub.android.ui.issue;
 
+import com.github.pockethub.android.R;
+import com.github.pockethub.android.core.issue.IssueStore;
 import com.github.pockethub.android.rx.AutoDisposeUtils;
 import com.github.pockethub.android.rx.RxProgress;
+import com.github.pockethub.android.ui.BaseActivity;
 import com.google.auto.factory.AutoFactory;
 import com.google.auto.factory.Provided;
 import com.meisolsson.githubsdk.model.Issue;
 import com.meisolsson.githubsdk.model.Milestone;
 import com.meisolsson.githubsdk.model.Repository;
-import com.github.pockethub.android.R;
-import com.github.pockethub.android.core.issue.IssueStore;
-import com.github.pockethub.android.ui.BaseActivity;
 import com.meisolsson.githubsdk.model.request.issue.IssueRequest;
-import com.uber.autodispose.AutoDispose;
-
-import javax.inject.Inject;
 
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.functions.Consumer;
@@ -91,7 +88,7 @@ public class EditMilestoneTask {
      */
     public EditMilestoneTask edit(Milestone milestone) {
         if (milestone != null) {
-            IssueRequest editedIssue = IssueRequest.builder().milestone(milestone.number()).build();
+            IssueRequest editedIssue = IssueRequest.builder().milestone(milestone.number().longValue()).build();
 
             store.editIssue(repositoryId, issueNumber, editedIssue)
                     .subscribeOn(Schedulers.io())
